requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==1.11.12
Flask==0.12.2
Flask-Cors==3.0.3
gunicorn==19.7.1
idna==2.6
image==1.5.20
iotop==0.6
itsdangerous==0.24
Jinja2==2.10
JPype1==0.6.2
MarkupSafe==1.0
meld3==1.0.2
MySQL-python==1.2.5
nltk==3.2.5
np==1.0.2
numpy==1.14.2
Pattern==2.6
Pillow==5.1.0
pycrypto==2.6.1
pycurl==7.43.0
pyliblzma==0.5.3
pysqlite==1.0.1
pytz==2018.4
requests==2.18.4
rpm-python==4.12.0.1
six==1.11.0
supervisor==3.3.4
urlgrabber==3.9.1
urllib3==1.22
Werkzeug==0.14.1
yum-metadata-parser==1.1.4
